Major Update
The website is now able to handle the positions that's been given to it appropriately with structured layout.

Support to result,json and config.yml has been solidified and reinforced.

Mobile Support has been halted due to time constraints.

Dynamic overlay has not been implemented currently thus limiting to only the given result lists future proof of this is planned and potentially be implemented before the deadline.